+++ l10n/af.js
@@ -1,0 +1,7 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "Authentication code" : "Verifikasiekode",
+    "Submit" : "Dien in"
+},
+"nplurals=2; plural=(n != 1);");