+++ l10n/ast.js
@@ -1,0 +1,8 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "Corréu",
+    "Authentication code" : "Códigu d'autenticación",
+    "Submit" : "Unviar"
+},
+"nplurals=2; plural=(n != 1);");