+++ l10n/de_DE.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "E-Mail",
+    "Proceed" : "Fortfahren",
+    "Authentication code" : "Authentifizierungscode",
+    "Submit" : "Übermitteln"
+},
+"nplurals=2; plural=(n != 1);");