+++ l10n/en_GB.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "E-mail",
+    "Proceed" : "Proceed",
+    "Authentication code" : "Authentication code",
+    "Submit" : "Submit"
+},
+"nplurals=2; plural=(n != 1);");