+++ l10n/eo.js
@@ -1,0 +1,7 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "Retpoŝto",
+    "Submit" : "Sendi"
+},
+"nplurals=2; plural=(n != 1);");