+++ l10n/gl.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "Correo-e",
+    "Proceed" : "Continuar",
+    "Authentication code" : "Código de autenticación",
+    "Submit" : "Enviar"
+},
+"nplurals=2; plural=(n != 1);");