+++ l10n/he.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "דוא״ל",
+    "Proceed" : "להמשיך",
+    "Authentication code" : "קוד אימות",
+    "Submit" : "שליחה"
+},
+"nplurals=3; plural=(n == 1 && n % 1 == 0) ? 0 : (n == 2 && n % 1 == 0) ? 1: 2;");