+++ l10n/id.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "E-mail",
+    "Proceed" : "Lanjutkan",
+    "Authentication code" : "Kode autentikasi",
+    "Submit" : "Kirim"
+},
+"nplurals=1; plural=0;");