+++ l10n/ka_GE.js
@@ -1,0 +1,8 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "ელ-ფოსტა",
+    "Authentication code" : "აუტენტიფიკაციის კოდი",
+    "Submit" : "გაგზავნა"
+},
+"nplurals=2; plural=(n!=1);");