+++ l10n/kab.js
@@ -1,0 +1,6 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "Proceed" : "Kemmel"
+},
+"nplurals=2; plural=(n != 1);");