+++ l10n/lo.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "ອີເມວ",
+    "Proceed" : "Proceed",
+    "Authentication code" : "Authentication code",
+    "Submit" : "ຕົກລົງ"
+},
+"nplurals=1; plural=0;");