+++ l10n/nb.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "E-post",
+    "Proceed" : "Fortsett",
+    "Authentication code" : "Autentiseringskode",
+    "Submit" : "Send inn"
+},
+"nplurals=2; plural=(n != 1);");