+++ l10n/nl.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "E-mail",
+    "Proceed" : "Doorgaan",
+    "Authentication code" : "Authenticatiecode",
+    "Submit" : "Indienen"
+},
+"nplurals=2; plural=(n != 1);");