+++ l10n/nn_NO.js
@@ -1,0 +1,7 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "E-post",
+    "Authentication code" : "Autentifikasjons kode"
+},
+"nplurals=2; plural=(n != 1);");