+++ l10n/sr.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "Е-пошта",
+    "Proceed" : "Настави",
+    "Authentication code" : "Код за проверу идентитета",
+    "Submit" : "Пошаљи"
+},
+"nplurals=3; plural=(n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2);");