+++ l10n/sw.js
@@ -1,0 +1,8 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "Proceed" : "Endelea",
+    "Authentication code" : "Msimbo wa uthibitishaji",
+    "Submit" : "Wasilisha"
+},
+"nplurals=2; plural=(n != 1);");