+++ l10n/th.js
@@ -1,0 +1,6 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "Submit" : "ส่ง"
+},
+"nplurals=1; plural=0;");