+++ l10n/zh_TW.js
@@ -1,0 +1,9 @@
+OC.L10N.register(
+    "twofactor_email",
+    {
+    "E-mail" : "電子郵件",
+    "Proceed" : "繼續",
+    "Authentication code" : "驗證碼",
+    "Submit" : "提交"
+},
+"nplurals=1; plural=0;");